Hoangphan

+++ routes/web.php
@@ -22,9 +22,6 @@ Route::get('/', function () {
     return view('front_end.home',compact('new_trans'));
 })->middleware('locale');
 
-// Route::get('/category', function () {
-//     return view('front_end.productlist');
-// })->middleware('locale');
 
 Route::get('/contact', function () {
     return view('front_end.contact');
@@ -32,44 +29,44 @@ Route::get('/contact', function () {
 
 Auth::routes();
 
-Route::name('admin.')->prefix('admin')->group(function () {
+Route::prefix('/admin')->group(function () {
     Route::get('/', function () {
         if (Auth::user()) {
-            return redirect('admin/dashboard');
+            return redirect('/admin/dashboard');
         }
         return view('auth.login');
-    })->name('login');
+    })->name('admin.login');
 
-    Route::get('dashboard', 'HomeController@index')->name('dashboard');
+    Route::get('dashboard', 'HomeController@index')->name('admin.dashboard');
 
-    Route::get('products/categories', 'ProductController@getCategory')->name('getCategory');
-    Route::get('products/get-deleted', 'ProductController@getDeleted')->name('products.getDeleted')->middleware('locale');
-    Route::get('products/restore/{id}', 'ProductController@restore')->name('products.restore');
-    Route::delete('products/force-delete/{id}', 'ProductController@forceDelete')->name('products.forceDelete')->middleware('locale');
+    Route::get('products/categories', 'ProductController@getCategory')->name('admin.getCategory');
+    Route::get('products/get-deleted', 'ProductController@getDeleted')->name('admin.products.getDeleted')->middleware('locale');
+    Route::get('products/restore/{id}', 'ProductController@restore')->name('admin.products.restore');
+    Route::delete('products/force-delete/{id}', 'ProductController@forceDelete')->name('admin.products.forceDelete')->middleware('locale');
 
     Route::resource('products', 'ProductController')->middleware('locale');
 
     Route::resource('category_new', 'Category_newController')->middleware('locale');
-    Route::get('category_new_deleted', 'Category_newController@show_deletad_at')->name('category_new.deleted_at')->middleware('locale');
-    Route::get('category_new_restore/{id}', 'Category_newController@restore')->name('category_new.restore')->middleware('locale');
-    Route::get('category_new_view/{id}', 'Category_newController@view_deleted_at')->name('category_new.view_deleted_at')->middleware('locale');
-    Route::delete('category_new_forceDelete/{id}', 'Category_newController@forceDelete')->name('category_new.forceDelete')->middleware('locale');
+    Route::get('category_new_deleted', 'Category_newController@show_deletad_at')->name('admin.category_new.deleted_at')->middleware('locale');
+    Route::get('category_new_restore/{id}', 'Category_newController@restore')->name('admin.category_new.restore')->middleware('locale');
+    Route::get('category_new_view/{id}', 'Category_newController@view_deleted_at')->name('admin.category_new.view_deleted_at')->middleware('locale');
+    Route::delete('category_new_forceDelete/{id}', 'Category_newController@forceDelete')->name('admin.category_new.forceDelete')->middleware('locale');
 
     Route::resource('new', 'NewController')->middleware('locale');
-    Route::get('new_deleted', 'NewController@show_deletad_at')->name('new.deleted_at')->middleware('locale');
-    Route::get('new_restore/{id}', 'NewController@restore')->name('new.restore')->middleware('locale');
-    Route::get('new_view/{id}', 'NewController@view_deleted_at')->name('new.view_deleted_at')->middleware('locale');
-    Route::delete('new_forceDelete/{id}', 'NewController@forceDelete')->name('new.forceDelete')->middleware('locale');
+    Route::get('new_deleted', 'NewController@show_deletad_at')->name('admin.new.deleted_at')->middleware('locale');
+    Route::get('new_restore/{id}', 'NewController@restore')->name('admin.new.restore')->middleware('locale');
+    Route::get('new_view/{id}', 'NewController@view_deleted_at')->name('admin.new.view_deleted_at')->middleware('locale');
+    Route::delete('new_forceDelete/{id}', 'NewController@forceDelete')->name('admin.new.forceDelete')->middleware('locale');
 
 
 // =======================================================================================================================>>>
 // =======================================================================================================================>>>
 
-    Route::get('shareholder/categories', 'Category_ShareHolder\ShareHolderController@getShareholderCategory')->name('getShareholderCategory');
+    Route::get('shareholder/categories', 'Category_ShareHolder\ShareHolderController@getShareholderCategory')->name('admin.getShareholderCategory');
 
-    Route::get('category-products/get-deleted', 'Category_Product\CategoryController@getDeleted')->name('category-products.getDelete')->middleware('locale');
-    Route::get('category-products/restore/{id}', 'Category_Product\CategoryController@restore')->name('category-products.restore')->middleware('locale');
-    Route::delete('category-products/force-delete/{id}', 'Category_Product\CategoryController@forceDelete')->name('category-products.forceDelete')->middleware('locale');
+    Route::get('category-products/get-deleted', 'Category_Product\CategoryController@getDeleted')->name('admin.category-products.getDelete')->middleware('locale');
+    Route::get('category-products/restore/{id}', 'Category_Product\CategoryController@restore')->name('admin.category-products.restore')->middleware('locale');
+    Route::delete('category-products/force-delete/{id}', 'Category_Product\CategoryController@forceDelete')->name('admin.category-products.forceDelete')->middleware('locale');
 
     Route::resource('category-products', 'Category_Product\CategoryController')->middleware('locale');
     Route::resource('category-shareholder', 'Category_ShareHolder\CateShareHolderController')->middleware('locale');
